Rearrange things a bit
