+++ ApplicationLibCode/FileInterface/RifReaderOpmCommon.h
@@ -59,37 +59,37 @@ public:
 
     std::vector<QDateTime> timeStepsOnFile( QString gridFileName );
 
-private:
-    void buildMetaData( RigEclipseCaseData* caseData, caf::ProgressInfo& progress );
-    bool importGrid( RigMainGrid* mainGrid, RigEclipseCaseData* caseData );
-    bool importActiveGrid( RigActiveCellGrid* mainGrid, RigEclipseCaseData* caseData );
-    void transferGeometry( Opm::EclIO::EGrid&  opmMainGrid,
-                           Opm::EclIO::EGrid&  opmGrid,
-                           RigMainGrid*        riMainGrid,
-                           RigGridBase*        riGrid,
-                           RigEclipseCaseData* caseData );
-    void transferActiveGeometry( Opm::EclIO::EGrid& opmMainGrid, RigActiveCellGrid* riMainGrid, RigEclipseCaseData* caseData );
+protected:
+    virtual bool importGrid( RigMainGrid* mainGrid, RigEclipseCaseData* caseData );
 
     void transferActiveCells( Opm::EclIO::EGrid&  opmGrid,
                               size_t              cellStartIndex,
                               RigEclipseCaseData* eclipseCaseData,
                               size_t              matrixActiveStartIndex,
                               size_t              fractureActiveStartIndex );
-
     void transferStaticNNCData( Opm::EclIO::EGrid& opmMainGrid, std::vector<Opm::EclIO::EGrid>& lgrGrids, RigMainGrid* mainGrid );
+
+    bool verifyActiveCellInfo( int activeSizeMat, int activeSizeFrac );
+    void updateActiveCellInfo( RigEclipseCaseData*             eclipseCaseData,
+                               Opm::EclIO::EGrid&              opmGrid,
+                               std::vector<Opm::EclIO::EGrid>& lgrGrids,
+                               RigMainGrid*                    mainGrid );
+
+private:
+    void buildMetaData( RigEclipseCaseData* caseData, caf::ProgressInfo& progress );
+
+    std::vector<RigEclipseTimeStepInfo> createFilteredTimeStepInfos();
+    std::vector<std::vector<int>>       readActiveCellInfoFromPorv( RigEclipseCaseData* eclipseCaseData, bool isDualPorosity );
+
+    void transferGeometry( Opm::EclIO::EGrid&  opmMainGrid,
+                           Opm::EclIO::EGrid&  opmGrid,
+                           RigMainGrid*        riMainGrid,
+                           RigGridBase*        riGrid,
+                           RigEclipseCaseData* caseData );
     void transferDynamicNNCData( RigMainGrid* mainGrid );
 
     void locateInitAndRestartFiles( QString gridFileName );
     void setupInitAndRestartAccess();
-
-    std::vector<RigEclipseTimeStepInfo> createFilteredTimeStepInfos();
-
-    bool                          verifyActiveCellInfo( int activeSizeMat, int activeSizeFrac );
-    std::vector<std::vector<int>> readActiveCellInfoFromPorv( RigEclipseCaseData* eclipseCaseData, bool isDualPorosity );
-    void                          updateActiveCellInfo( RigEclipseCaseData*             eclipseCaseData,
-                                                        Opm::EclIO::EGrid&              opmGrid,
-                                                        std::vector<Opm::EclIO::EGrid>& lgrGrids,
-                                                        RigMainGrid*                    mainGrid );
 
     struct TimeDataFile
     {
@@ -102,22 +102,22 @@ private:
 
     std::vector<TimeDataFile> readTimeSteps();
 
-private:
+protected:
     enum class ActiveType
     {
         ACTIVE_MATRIX_VALUE   = 1,
         ACTIVE_FRACTURE_VALUE = 2
     };
 
-    std::string m_gridFileName;
-    std::string m_initFileName;
-    std::string m_restartFileName;
-    int         m_gridUnit;
+    std::string              m_gridFileName;
+    int                      m_gridUnit;
+    std::vector<std::string> m_gridNames;
 
     RigEclipseCaseData* m_eclipseCaseData;
 
+private:
+    std::string                        m_initFileName;
+    std::string                        m_restartFileName;
     std::unique_ptr<Opm::EclIO::ERst>  m_restartFile;
     std::unique_ptr<Opm::EclIO::EInit> m_initFile;
-
-    std::vector<std::string> m_gridNames;
 };
